Cross-platform visual e2e testing enabled
