+++ DotnetCore/Sauce.Demo/Core.BestPractices.Web/Tests/AllTestsBase.cs
@@ -14,20 +14,5 @@ namespace Core.BestPractices.Web.Tests
         public string SauceAccessKey;
         public Dictionary<string, object> SauceOptions;
         public string ScreenerApiKey;
-
-        [SetUp]
-        public void Setup()
-        {
-            SauceUserName = Environment.GetEnvironmentVariable("SAUCE_USERNAME", EnvironmentVariableTarget.User);
-            SauceAccessKey = Environment.GetEnvironmentVariable("SAUCE_ACCESS_KEY", EnvironmentVariableTarget.User);
-            ScreenerApiKey = Environment.GetEnvironmentVariable("SCREENER_API_KEY", EnvironmentVariableTarget.User);
-            SauceOptions = new Dictionary<string, object>
-            {
-                ["username"] = SauceUserName,
-                ["accessKey"] = SauceAccessKey,
-                ["name"] = TestContext.CurrentContext.Test.Name
-            };
-        }
-
     }
 }
