+++ DotnetCore/Sauce.Demo/Core.BestPractices.Web/Tests/WebTestsBase.cs
@@ -1,6 +1,8 @@
 using NUnit.Framework;
 using NUnit.Framework.Interfaces;
 using OpenQA.Selenium;
+using System;
+using System.Collections.Generic;
 
 namespace Core.BestPractices.Web.Tests
 {
@@ -14,6 +16,19 @@ namespace Core.BestPractices.Web.Tests
                 return;
             ExecuteSauceCleanupSteps();
             Driver.Quit();
+        }
+        [SetUp]
+        public void Setup()
+        {
+            SauceUserName = Environment.GetEnvironmentVariable("SAUCE_USERNAME", EnvironmentVariableTarget.User);
+            SauceAccessKey = Environment.GetEnvironmentVariable("SAUCE_ACCESS_KEY", EnvironmentVariableTarget.User);
+            ScreenerApiKey = Environment.GetEnvironmentVariable("SCREENER_API_KEY", EnvironmentVariableTarget.User);
+            SauceOptions = new Dictionary<string, object>
+            {
+                ["username"] = SauceUserName,
+                ["accessKey"] = SauceAccessKey,
+                ["name"] = TestContext.CurrentContext.Test.Name
+            };
         }
         private void ExecuteSauceCleanupSteps()
         {
